model_trainer/requirements.txt
# Python dependencies for feedback extraction and PEFT/LoRA training
python-dotenv
pymongo
datasets
transformers
torch
accelerate
peft
